# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vokal Director API",
    "description": "Контракт между фронтендом и сервером Vokal Director.\n\n**Что здесь важно знать до чтения адресов.**\n\n* Большая часть адресов отвечает `501`. Это не заготовка на будущее: схема\n  запроса и ответа настоящая и полная, реализации за адресом еще нет. В\n  `details.missing` сказано, чего именно не хватает. Заглушек, возвращающих\n  правдоподобные данные, в API нет — мок в ответе сервера хуже отказа.\n* Ошибка всегда выглядит одинаково: `{\"error\": {\"code\", \"message\",\n  \"requestId\", \"details\"}}`.\n* Имена полей — camelCase. Идентификаторы — непрозрачные строки.\n* Незнакомое поле в теле запроса — ошибка `422`, а не тихо отброшенное\n  значение.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/review-issues": {
      "get": {
        "tags": [
          "проверка"
        ],
        "summary": "Места, которые стоит проверить",
        "description": "Отбор по статусу необязателен. У песни без разбора список отвечает `501`: пустой список читался бы как «все в порядке».",
        "operationId": "list_issues_api_projects__project_id__review_issues_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReviewStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Отбор по статусу проверки",
              "title": "Status"
            },
            "description": "Отбор по статусу проверки"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewIssueListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Данные запроса не прошли проверку схемы.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/review-issues/{issue_id}": {
      "patch": {
        "tags": [
          "проверка"
        ],
        "summary": "Сменить статус места",
        "description": "Пять статусов различают «исправлено» и «принято для репетиции»: во втором случае место осталось спорным, но играть договорились так.",
        "operationId": "update_issue_status_api_projects__project_id__review_issues__issue_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          },
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор места проверки",
              "title": "Issue Id"
            },
            "description": "Идентификатор места проверки"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewIssueStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewIssueOut"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Состояние объекта изменилось: обновите данные и повторите.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Данные запроса не прошли проверку схемы.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/review-issues/{issue_id}/comments": {
      "post": {
        "tags": [
          "проверка"
        ],
        "summary": "Свой комментарий к месту",
        "description": "Статус места при этом не меняется: комментарий — не проверка.",
        "operationId": "add_comment_api_projects__project_id__review_issues__issue_id__comments_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          },
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор места проверки",
              "title": "Issue Id"
            },
            "description": "Идентификатор места проверки"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCommentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewCommentOut"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Данные запроса не прошли проверку схемы.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Машинный код для ветвления на клиенте"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Текст для показа пользователю"
          },
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid",
            "description": "Идентификатор запроса: по нему поддержка ищет след в логах"
          },
          "details": {
            "title": "Details",
            "description": "Подробности, зависят от кода. Для validation_error — {fields: [{field, reason}]}, для not_implemented — {endpoint, missing, docs}."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "description": "Тело ошибки. Форма задана `app/core/errors.py` и одинакова для всего API."
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "Конверт ошибки: любой ответ с кодом 4xx/5xx выглядит так."
      },
...
      "ReviewCommentCreate": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text"
          },
          "author": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Author",
            "description": "Подпись автора. Пусто — подставляется владелец запроса; чужое имя подставить нельзя."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ReviewCommentCreate",
        "description": "Свой комментарий к месту проверки."
      },
      "ReviewCommentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "issueId": {
            "type": "string",
            "title": "Issueid"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "issueId",
          "author",
          "text",
          "createdAt"
        ],
        "title": "ReviewCommentOut",
        "description": "Комментарий человека к месту проверки."
      },
      "ReviewIssueListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewIssueOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ReviewIssueListResponse"
      },
      "ReviewIssueOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "sectionId": {
            "type": "string",
            "title": "Sectionid"
          },
          "bar": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Bar"
          },
          "part": {
            "type": "string",
            "title": "Part"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "Почему место попало в проверку"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence",
            "description": "Уверенность: 0..1. Пусто — рассчитать ее нечем, причина в confidenceNote."
          },
          "confidenceNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidencenote",
            "description": "Почему уверенности нет. Заполнено ровно тогда, когда пусто число."
          },
          "comments": {
            "items": {
              "$ref": "#/components/schemas/ReviewCommentOut"
            },
            "type": "array",
            "title": "Comments"
          },
          "confidencePercent": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidencepercent",
            "description": "Уверенность в процентах: то же число, что и рядом с полосой на экране",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "sectionId",
          "bar",
          "part",
          "reason",
          "status",
          "confidencePercent"
        ],
        "title": "ReviewIssueOut",
        "description": "Место, которое стоит проверить руками."
      },
      "ReviewIssueStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "Необязательное пояснение: сохраняется как комментарий к месту"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ReviewIssueStatusUpdate",
        "description": "Смена статуса места проверки."
      },
      "ReviewStatus": {
        "type": "string",
        "enum": [
          "needs_review",
          "checked",
          "fixed",
          "uncertain",
          "accepted_for_rehearsal"
        ],
        "title": "ReviewStatus"
      },
...
    }
  }
}